Extracted Footer to separate common component
