+++ src/components/Space/SpaceResultItem.js
