+++ src/containers/SearchResultsContainer.spec.js
@@ -1,16 +1,16 @@
 import React from 'react';
 import { shallow } from 'enzyme';
 
-import { SearchResultsPage } from './SearchResultsPage';
-import SearchResults from '../components/SearchResults';
+import SearchResultsContainer from './SearchResultsContainer';
+import SearchResults from '../components/Search/SearchResults';
 
-describe('<SearchResultsPage />', () => {
+describe('<SearchResultsContainer />', () => {
     it('should contain <SearchResults />', () => {
         const actions = {
             searchSpaces: () => {}
         };
 
-        const wrapper = shallow(<SearchResultsPage actions={actions} />);
+        const wrapper = shallow(<SearchResultsContainer actions={actions} />);
         expect(wrapper.find(SearchResults).length).toEqual(1);
     });
 });
